[Environments] Add the concept of environments
Add the concept of environments using the Xcode schemes

Currently there are three environments: DEV, QA and PROD

The idea behind this is that information suchs as Service Keys/Infos
such as (Parse, Anaylitics, Crashlogs, etc) and Base URL for backend
communitication setup should be per environment.

This setup generates a enviornment.plist files in the environments
folder before every build which will be based on the current
environment, reason why it should be in the gitingore. The prefixed
header file is also generated before every build, so it is also in the
.gitingore file

Update README file regarding the environment

Add parser information into the environment.plist file that should be
used to configure the parse library